Moved UppyUploadFile to the main folder
