+++ Clients/src/presentation/components/Inputs/FileUpload/getStyles.ts
